feat(ux): :wrench: fully implemented configHandler
Configuration handler in seperate file has been added to improve configuration ux.
